Read the build mode like the other options.
(incredi-only) : Command removed.
(incredi--read-info) : Does not receive the mode anymore, but reads it.
The output plist does not contain the executable anymore, but contains
the mode and the conditions outside manages the executable.
(incredi--build-command) : Does not receive the mode either and the
pcase uses the :mode in the plist.
(incredi--build-internal) : Receives now the plist that must be
computed outside now.
(incredi-build) : Calls the incredi--build-internal passing the plist.