Add a second parse to get project configs.
This allows to get the available config (debug, release) and the architectures.
This extra parsing does not add extra cost because the Global section
is restricted to a file portion at the end.
I am not sure on how the sln format is properly set and if the Global is
at the end; so I add an extra goto-pos point-min.